backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
psycopg2-binary
apscheduler
pydantic
gunicorn